pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-cors>=6.0.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "requests>=2.32.4",
    "pillow>=11.2.1",
    "flask-swagger-ui>=5.21.0",
    "google-generativeai>=0.8.5",
    "flask-jwt-extended>=4.7.1",
    "qrcode>=8.2",
]
